refactor: 폴더 구조 및 이름 변경
