+++ src/components/dashboard/layout/dashboard-frame/index.tsx
@@ -1,12 +1,14 @@
-import DashboardHeader from '@components/dashboard-header';
-import Sidebar from '@components/sidebar';
+import { Outlet } from 'react-router-dom';
 import { DashboardLayout } from './index.style';
+import Sidebar from '@components/sidebar';
+import DashboardHeader from '@components/dashboard/layout/dashboard-header';
 
 const DashboardFrame: React.FC = () => {
   return (
     <DashboardLayout>
       <DashboardHeader />
       <Sidebar />
+      <Outlet />
     </DashboardLayout>
   );
 };
