+++ src/components/dashboard/modals/register-modal/index.tsx
@@ -1,7 +1,8 @@
 import React, { useState, useCallback } from 'react';
 import ReactDOM from 'react-dom';
-import { Label, ModalContent, Overlay, InputContainer, InputWrapper, InputField, CloseButton } from './index.style';
-import { ReactComponent as CloseIcon } from '../../../assets/CloseIcon.svg';
+import ButtonGroup from '../../button';
+import { Label, ModalContent, Overlay, ButtonWrapper, InputContainer, InputWrapper, InputField, CloseButton } from './index.style';
+import { ReactComponent as CloseIcon } from '../../../../assets/CloseIcon.svg';
 
 const useModal = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -18,7 +19,7 @@ interface ModalProps {
   fields: { label: string; placeholder: string }[];
 }
 
-const KeychainReadModal: React.FC<ModalProps> = ({ isOpen, onClose, fields }) => {
+const RegisterModal: React.FC<ModalProps> = ({ isOpen, onClose, fields }) => {
   if (!isOpen) return null;
 
   const [inputValues, setInputValues] = useState<string[]>(fields.map(() => ''));
@@ -34,7 +35,11 @@ const KeychainReadModal: React.FC<ModalProps> = ({ isOpen, onClose, fields }) =>
   return ReactDOM.createPortal(
     <Overlay onClick={onClose}>
       <ModalContent onClick={(e) => e.stopPropagation()} style={{ margin: '24px' }}>
-        <CloseButton onClick={onClose}>
+        <CloseButton
+          onClick={() => {
+            onClose();
+          }}
+        >
           <CloseIcon />
         </CloseButton>
 
@@ -42,10 +47,14 @@ const KeychainReadModal: React.FC<ModalProps> = ({ isOpen, onClose, fields }) =>
           {fields.map((field, index) => (
             <InputContainer key={index}>
               <Label>{field.label}</Label>
-              <InputField placeholder={field.placeholder} value={inputValues[index]} onChange={(e) => handleInputChange(index, e.target.value)} readOnly />
+              <InputField placeholder={field.placeholder} value={inputValues[index]} onChange={(e) => handleInputChange(index, e.target.value)} />
             </InputContainer>
           ))}
         </InputWrapper>
+
+        <ButtonWrapper>
+          <ButtonGroup inputValues={inputValues} />
+        </ButtonWrapper>
       </ModalContent>
     </Overlay>,
     document.body
@@ -54,4 +63,4 @@ const KeychainReadModal: React.FC<ModalProps> = ({ isOpen, onClose, fields }) =>
 
 export { useModal };
 export type { ModalProps };
-export default KeychainReadModal;
+export default RegisterModal;
